terminando o CRUD de categorias

+++ ms-tcAtelie-auth/src/main/java/tcatelie/microservice/auth/service/CategoriaService.java
@@ -8,11 +8,15 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
+import tcatelie.microservice.auth.dto.filter.ProdutoFiltroDTO;
 import tcatelie.microservice.auth.dto.request.CategoriaRequestDTO;
 import tcatelie.microservice.auth.dto.response.CategoriaResponseDTO;
 import tcatelie.microservice.auth.mapper.CategoriaMapper;
 import tcatelie.microservice.auth.model.Categoria;
+import tcatelie.microservice.auth.model.Produto;
 import tcatelie.microservice.auth.repository.CategoriaRepository;
+import tcatelie.microservice.auth.repository.ProdutoRepository;
+import tcatelie.microservice.auth.specification.ProdutoSpecification;
 
 import java.util.List;
 import java.util.Optional;
@@ -23,6 +27,7 @@ public class CategoriaService {
 
     private final CategoriaRepository repository;
     private final CategoriaMapper mapper;
+    private final ProdutoRepository produtoRepository;
 
     public CategoriaResponseDTO cadastrarCategoria(CategoriaRequestDTO requestDTO) {
         validarRequest(requestDTO);
@@ -73,11 +78,9 @@ public class CategoriaService {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         }
 
-        Categoria categoriaEntidade = mapper.toCategoria(categoriaRequestDTO);
-        categoriaEntidade.setIdCategoria(id);
+        Categoria categoriaEntidade = findById(id);
         categoriaEntidade.setNomeCategoria(categoriaRequestDTO.getNome());
         Categoria categoriaSalva = repository.save(categoriaEntidade);
-
         return mapper.toCategoriaResponse(categoriaSalva);
     }
 
@@ -85,6 +88,12 @@ public class CategoriaService {
         if(repository.findById(id).isEmpty()){
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         }
-        repository.deleteById(id);
+        List<Produto> listaProdutos = produtoRepository.findAll(ProdutoSpecification.filtrar(ProdutoFiltroDTO.builder().idCategoria(id).build()));
+
+        if(listaProdutos.isEmpty()){
+            repository.deleteById(id);
+        }
+
+
     }
 }
